fix: 修复循环依赖，优化代码结构
