+++ services/command/gpt_chat_service.go
@@ -8,7 +8,7 @@ import (
 	"github.com/baiyz0825/corp-webot/config"
 	"github.com/baiyz0825/corp-webot/dao"
 	"github.com/baiyz0825/corp-webot/model"
-	"github.com/baiyz0825/corp-webot/services/impl"
+	"github.com/baiyz0825/corp-webot/services/wx"
 	"github.com/baiyz0825/corp-webot/to"
 	"github.com/baiyz0825/corp-webot/utils/openaiutils"
 	"github.com/baiyz0825/corp-webot/utils/xlog"
@@ -37,9 +37,9 @@ func (c GPTChatCommand) Exec(userData to.MsgContent) bool {
 	mode := config.GetSystemConf().MsgMode
 	switch mode {
 	case "markdown":
-		return impl.SendToWxByMarkdown(userData, respOpenAI)
+		return wx.SendToWxByMarkdown(userData, respOpenAI)
 	case "text":
-		return impl.SendToWxByText(userData, respOpenAI)
+		return wx.SendToWxByText(userData, respOpenAI)
 	default:
 		return false
 	}
