+++ services/command/gpt_image_service.go
